Create Checkboxes as Radio Buttons
I do not claim the majority of this code, it was found here -> https://spreadsheet.dev/radio-buttons-in-google-sheets-using-apps-script - and edited to meet the needs of the project as necessary. This is the edited code. All edits done by me, but I did not make the original code. The edited code is designed for use with a specific Google Sheets spreadsheet, but can be easily edited based on the original code found at the link above or my edited version depending on the needs of the script in question.

The purpose of this code is to simulate radio buttons in a Google Sheets spreadsheet, as Google Sheets does not natively support radio buttons as of now. This is accomplished by using checkboxes on the spreadsheet itself, and then looking for edit triggers, checking if the trigger was the checking of a checkbox, then unchecking all other checkboxes. This code is set up to work with 2 checkboxes at a time in either a row or column, but can easily be expanded to many checkboxes, as the original code (see link) allows for functionality with an entire column.