Create Data Filter Function
This function is designed to emulate Microsoft Excel and Visual Basic's AutoFilter function, which does not have a direct equivalent in Google Sheets or Google Apps Script. This function was designed and written by me in order to perform the same functionality in Sheets as in Excel, but in a slightly more limited scope. This code is designed with a specific spreadsheet in mind, but can be tweaked and edited to perform similar functionality within other spreadsheets if necessary. The function is also  specifically designed in such a way that it will filter an ENTIRE sheet, from A1 to whatever cell is specified. This can be edited, but that is the default functionality. The function supports cell ranges up to column ZZ. Double-letter columns are untested, but should work correctly.

The parameter 'sheet' is anticipated to be a specific Sheet, obtained by using SpreadsheetApp.getActiveSpreadsheet() class. The parameter 'cell' is anticipated to be a string value of a cell's A1-format designation (Example: "A75"). 'cell' is the END of the data range to be filtered, starting from A1. The parameter 'search' is anticipated to be a value that the function will search for within the data range A1 to 'cell.' Only data matching 'search' will be shown in the sheet after running the function.