Create Plot File Data to Spreadsheet Function
This function is designed with a specific spreadsheet and output file format, but certain aspects of the function can be edited in order to work with other formats of files and spreadsheets. The output file in question is created by an automatic, proprietary program used by the WV DEP that prints its output files in a specific text format that allows for this function to exploit patterns in the formatting in order to accurately plot the two distinct halves of the output data. The first half, being a header, is printed within a single column as it is mostly all text. The second half, being a table of data values, needs to be plotted as such, with values being in separated cells. This function is designed with these requirements in mind. Due to security concerns, I cannot provide an example of the output file format that the function is designed around, but this code should work with other file formats with some tweaking. The file being plotted is expected to be within the user's Google Drive. Upon testing, the file can be in or out of a folder with no significant effect on performance. This function was created due to a need for a way to plot file data to a spreadsheet as Google Sheets lacks an internal function to do so like Excel has.